backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django4-background-tasks==1.2.7
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
drf-spectacular==0.26.4
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
oauthlib==3.2.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
